+++ zerobug/scripts/main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# template 23
+# template 24
 """
 Python supplemental features
 ----------------------------
@@ -20,39 +20,84 @@ virtual environment.
 import os
 import os.path as pth
 import sys
-import pkg_resources
+# import pkg_resources
 import gzip
 import shutil
 
 
-__version__ = "2.0.16"
+__version__ = "2.0.18"
+
+
+def get_pypi_info(pkgname):
+    pypi_metadata = {
+        "name": pkgname, "version": False, "where": "pylib",
+        "long_description": __doc__,
+    }
+    if sys.version_info[0] == 2:
+        import pkg_resources
+        try:
+            pypi_metadata["version"] = pkg_resources.get_distribution(pkgname).version
+        except BaseException:
+            pass
+    else:
+        if sys.version_info < (3, 8):
+            import importlib_metadata as metadata
+        else:
+            from importlib import metadata
+        try:
+            pypi_metadata["version"] = metadata.version(pkgname)
+            # for ln in metadata.metadata(pkgname).items():
+            #     if ln[0] == "Description":
+            #         pypi_metadata["long_description"] = ln[1]
+            #         break
+        except BaseException:
+            pass
+    return pypi_metadata
 
 
 def fake_setup(**kwargs):
-    globals()["setup_args"] = kwargs
+    # globals()["setup_args"] = kwargs
+    globals()["pypi_metadata"] = kwargs
 
 
-def read_setup():
-    setup_info = pth.abspath(pth.join(pth.dirname(__file__), "setup.info"))
-    if not pth.isfile(setup_info):
-        setup_info = pth.abspath(
-            pth.join(pth.dirname(__file__), "..", "setup.py")
-        )
-    setup_args = {}
+def get_metadata():
+    pkgpath = pth.abspath(pth.join(pth.dirname(pth.realpath(__file__)), "..", ".."))
+    setup_info = pth.join(pkgpath, "setup.py")
+    pypi_metadata = {}
     if pth.isfile(setup_info):
         with open(setup_info, "r") as fd:
             exec(fd.read().replace("setup(", "fake_setup("))
-            setup_args = globals()["setup_args"]
-    else:
-        print("Not internal configuration file found!")
-    setup_args["setup"] = setup_info
-    try:
-        pkg = pkg_resources.get_distribution(__package__.split(".")[0])
-        setup_args["name"] = pkg.key
-        setup_args["version"] = pkg.version
-    except BaseException:
-        pass
-    return setup_args
+            pypi_metadata = globals()["pypi_metadata"]
+    if not pypi_metadata:
+        return get_pypi_info(__package__.split(".")[0])
+    # readme = pth.join(pkgpath, "README.rst")
+    # with open(readme, "r") as fd:
+    #     pypi_metadata["long_description"] = fd.read()
+    pypi_metadata["where"] = "local"
+    return pypi_metadata
+
+
+# def read_setup():
+#     setup_info = pth.abspath(pth.join(pth.dirname(__file__), "setup.info"))
+#     if not pth.isfile(setup_info):
+#         setup_info = pth.abspath(
+#             pth.join(pth.dirname(__file__), "..", "setup.py")
+#         )
+#     setup_args = {}
+#     if pth.isfile(setup_info):
+#         with open(setup_info, "r") as fd:
+#             exec(fd.read().replace("setup(", "fake_setup("))
+#             setup_args = globals()["setup_args"]
+#     else:
+#         print("Not internal configuration file found!")
+#     setup_args["setup"] = setup_info
+#     try:
+#         pkg = pkg_resources.get_distribution(__package__.split(".")[0])
+#         setup_args["name"] = pkg.key
+#         setup_args["version"] = pkg.version
+#     except BaseException:
+#         pass
+#     return setup_args
 
 
 def get_pypi_paths():
@@ -80,8 +125,8 @@ def get_pypi_paths():
     return pkgpath, bin_path, lib_path
 
 
-def copy_pkg_data(setup_args, verbose):
-    if setup_args.get("package_data"):
+def copy_pkg_data(pypi_metadata, verbose):
+    if pypi_metadata.get("package_data"):
         pkgpath, bin_path, lib_path = get_pypi_paths()
         if bin_path:
             # TODO> compatibility mode
@@ -91,8 +136,8 @@ def copy_pkg_data(setup_args, verbose):
             man_path = pth.join(bin_path, "man", "man8")
             if not pth.isdir(man_path):
                 man_path = ""
-            for pkg in setup_args["package_data"].keys():
-                for fn in setup_args["package_data"][pkg]:
+            for pkg in pypi_metadata["package_data"].keys():
+                for fn in pypi_metadata["package_data"][pkg]:
                     base = pth.basename(fn)
                     if base in ("setup.info", "*"):
                         continue
@@ -158,26 +203,31 @@ def main(cli_args=None):
             action = arg
         elif arg == "-v":
             verbose = True
-    setup_args = read_setup()
+    # setup_args = read_setup()
+    pypi_metadata = get_metadata()
     if action == "-h":
         print(
             "%s [-h][-H][--help][-V][--version][-C][--copy-pkg-data]"
-            % setup_args["name"]
+            # % setup_args["name"]
+            % pypi_metadata["name"]
         )
     elif action in ("-V", "--version"):
-        if setup_args["version"] == __version__:
-            print(setup_args["version"])
+        # if setup_args["version"] == __version__:
+        #     print(setup_args["version"])
+        # else:
+        #     print("Version mismatch %s/%s" % (setup_args["version"], __version__))
+        if pypi_metadata["version"] == __version__:
+            print(pypi_metadata["version"])
         else:
-            print("Version mismatch %s/%s" % (setup_args["version"], __version__))
+            print("Version mismatch %s/%s" % (pypi_metadata["version"], __version__))
     elif action in ("-H", "--help"):
-        for text in __doc__.split("\n"):
+        # for text in __doc__.split("\n"):
+        #     print(text)
+        for text in pypi_metadata["long_description"].split("\n"):
             print(text)
     elif action in ("-C", "--copy-pkg-data"):
-        copy_pkg_data(setup_args, verbose)
+        # copy_pkg_data(setup_args, verbose)
+        copy_pkg_data(pypi_metadata, verbose)
     return 0
-
-
-
-